BUG: don't try to delete shared object files
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=6915
git-svn-id: http://svn.slicer.org/Slicer4/trunk@6915 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Base/Logic/vtkSlicerApplicationLogic.cxx
@@ -1244,7 +1244,15 @@ void vtkSlicerApplicationLogic::ProcessReadNodeData(ReadDataRequest& req)
     if (req.GetDeleteFile())
       {
       int removed;
-      removed = itksys::SystemTools::RemoveFile( req.GetFilename().c_str() );
+      // is it a shared memory location?
+      if (req.GetFilename().find("slicer:0x") != std::string::npos)
+        {
+        removed = 1;
+        }
+      else
+        {
+        removed = itksys::SystemTools::RemoveFile( req.GetFilename().c_str() );
+        }
       if (!removed)
         {
         std::stringstream information;
